Add analyzer rulesets and XML Documentation file
A global analyzer ruleset, Rezare.rSite.ruleset, has been added to
the solution. This is stored in the solution directory
'solution items'. Project specific rulesets have been added to the
Domain and Application projects. The global ruleset is not applied
to a project until it is included with a project specific ruleset.

StyleCop.Analyzer requires an XML Documentation Output file to be
included with the project.