Update code to allow deployment on AWS
The API deployment is done using Publish to Elastic Beanstalk
through Visual Studio.

The Angular site is stored in its own S3 bucket that is staticly
hosted. But the goal is to use AWS CodePipelines to automatically
deploy a new copy on commit. To allow this a buildspec.yml file has
been added.