Update solution to use nullable reference types
Enabling nullable reference types required making a number of
changes in the solution. This involved:

 - Suppressing rule CA1707 in Tests and CA1062 in Domain

 - Updating NuGet packages (not explicitly needed, but done anyway)

 - Fixing failing rules

 - Updating Startup.cs for .net core 3

 - Setting some reference types to be nullable

 - Updating ValueObject and its tests

There is a known issue with .net core 3 in which reference values
are not detected as not being null. This should be fixed in .net
core 3.1 and Visual Studio 16.4. The following lines have a
null-conditional operator or a null-forgiving operator:

 - ValueObject.cs lines 51, 73 and 80

 - ValueObjectTests_Equals.cs lines 18 and 32